backend/requirements/default.txt
aioboto3==15.1.0
aiohttp==3.12.14
alembic==1.10.4
asyncpg==0.30.0
atlassian-python-api==3.41.16
beautifulsoup4==4.12.3
boto3==1.39.11
celery==5.5.1
chardet==5.2.0
chonkie==1.0.10
dask==2023.8.1
ddtrace==3.10.0
discord.py==2.4.0
distributed==2023.8.1
fastapi==0.116.1
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==5.0.0
fastmcp==2.11.3
filelock==3.15.4
google-api-python-client==2.86.0
google-cloud-aiplatform==1.121.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
# GPT4All library has issues running on Macs and python:3.11.4-slim-bookworm
# will reintroduce this when library version catches up
# gpt4all==2.0.2
httpcore==1.0.9
httpx[http2]==0.28.1
httpx-oauth==0.15.1
huggingface-hub==0.35.3
inflection==0.5.1
jira==3.10.5
jsonref==1.1.0
trafilatura==1.12.2
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langgraph==0.2.72
langgraph-checkpoint==2.0.13
langgraph-sdk==0.1.44
lazy_imports==1.0.1
litellm==1.79.0
lxml==5.3.0
Mako==1.2.4
markitdown[pdf, docx, pptx, xlsx, xls]==0.1.2
mcp[cli]==1.12.4
msal==1.28.0
nltk==3.9.1
Office365-REST-Python-Client==2.5.9
oauthlib==3.2.2
openai==2.6.1
openpyxl==3.0.10
passlib==1.7.4
playwright==1.55.0
psutil==5.9.5
psycopg2-binary==2.9.9
puremagic==1.28
pyairtable==3.0.1
pycryptodome==3.19.1
pydantic==2.11.7
PyGithub==2.5.0
pympler==1.1
python-dateutil==2.8.2
python-gitlab==5.6.0
python-pptx==0.6.23
pypdf==6.1.3
pytest-mock==3.12.0
pytest-playwright==0.7.0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pysmb==1.2.9
pywikibot==9.0.0
redis==5.0.8
requests==2.32.5
requests-oauthlib==1.3.1
retry==0.9.2  # This pulls in py which is in CVE-2022-42969, must remove py from image
rfc3986==1.5.0
simple-salesforce==1.12.6
slack-sdk==3.20.2
SQLAlchemy[mypy]==2.0.15
starlette==0.47.2
supervisor==4.2.5
RapidFuzz==3.13.0
tiktoken==0.7.0
timeago==1.0.16
types-openpyxl==3.0.4.7
unstructured==0.15.1
unstructured-client==0.25.4
uvicorn==0.35.0
zulip==0.8.2
hubspot-api-client==8.1.0
asana==5.0.8
dropbox==11.36.2
boto3-stubs[s3]==1.39.11
shapely==2.0.6
stripe==10.12.0
urllib3==2.5.0
mistune==0.8.4
sentry-sdk==2.14.0
prometheus_client==0.21.0
fastapi-limiter==0.1.6
prometheus_fastapi_instrumentator==7.1.0
sendgrid==6.11.0
voyageai==0.2.3
cohere==5.6.1
exa_py==1.15.4
braintrust[openai-agents]==0.2.6
braintrust-langchain==0.0.4
openai-agents==0.4.2
langfuse==3.10.0
nest_asyncio==1.6.0
openinference-instrumentation-openai-agents==1.3.0
opentelemetry-proto==1.38.0
python3-saml==1.15.0
xmlsec==1.3.14
